+++ jest.config.js
@@ -31,7 +31,7 @@ module.exports = {
     '<rootDir>/src/**/*.jsx',
     '<rootDir>/src/**/*.tsx',
   ],
-  coveragePathIgnorePatterns: ['/node_modules/', '.d.ts$', '/src/[boot|i18n|router]/*', '/src/App.vue', '/src/stores/index.ts'],
+  coveragePathIgnorePatterns: ['/node_modules/', '.d.ts$', '/src/[boot|router]/*', '/src/App.vue', '/src/stores/index.ts'],
   coverageReporters: ['text', 'text-summary', 'cobertura'],
   coverageThreshold: {
     global: {
@@ -44,9 +44,7 @@ module.exports = {
   testMatch: [
     // Matches tests in any subfolder of 'src' or into 'test/jest/__tests__'
     // Matches all files with extension 'js', 'jsx', 'ts' and 'tsx'
-    // '<rootDir>/test/**/*.(spec|test).+(ts|js)?(x)', // TODO: temporary disabled until 0.2.0, uncomment then
-    '<rootDir>/test/composables/*.(spec|test).+(ts|js)?(x)',
-    '<rootDir>/test/stores/*.(spec|test).+(ts|js)?(x)',
+    '<rootDir>/test/**/*.(spec|test).+(ts|js)?(x)',
   ],
   // Extension-less imports of components are resolved to .ts files by TS,
   //  grating correct type-checking in test files.