+++ quasar.config.js
@@ -31,6 +31,7 @@ module.exports = configure((/* ctx */) => ({
   // --> boot files are part of "main.js"
   // https://v2.quasar.dev/quasar-cli-vite/boot-files
   boot: [
+    'config',
     'i18n',
   ],